moved env config to separate file and fix variables
