+++ src/main/resources/reference.conf
@@ -1,13 +1,13 @@
 couchbase-environment {
   
   # The connect timeout is used when a Bucket is opened and if not overridden by a custom timeout. It does not affect asynchronous operations. If you feel the urge to change this value to something higher, there is a good chance that your network is not properly set up. Opening a bucket should in practice not take longer than a second on a reasonably fast network.
-  connect-timeout = "5000ms"
+  connect-timeout = "5000 millis"
   
   # The amount of time the SDK will wait on the socket connect until an error is raised and handled.
-  socket-connect-timeout = "1000ms"
+  socket-connect-timeout = "1000 millis"
 
   # how often to retry to open the bucket if it's throwing an error
-  open-bucket-retry-timeout = "10s"
+  open-bucket-retry-timeout = "10 seconds"
 }
 
 couchbase-journal {
