+++ src/main/scala/akka/persistence/couchbase/CouchbasePluginConfig.scala
@@ -10,7 +10,6 @@ import com.couchbase.client.java.view.Stale
 import com.typesafe.config.Config
 
 import scala.concurrent.duration.FiniteDuration
-import scala.concurrent.duration.Duration
 
 trait CouchbasePluginConfig {
 
@@ -64,32 +63,6 @@ abstract class DefaultCouchbasePluginConfig(config: Config) extends CouchbasePlu
         cluster.openBucket(bucketName)
     }
 
-  }
-}
-
-trait CouchbaseEnvironmentConfig extends CouchbasePluginConfig {
-
-  def connectTimeout: FiniteDuration
-  
-  def socketConnectTimeout: FiniteDuration
-
-  def openBucketRetryTimeout: FiniteDuration
-}
-
-class DefaultCouchbaseEnvironmentConfig(config: Config)
-  extends DefaultCouchbasePluginConfig(config)
-    with CouchbaseEnvironmentConfig {
-  
-  override val connectTimeout = FiniteDuration(config.getDuration("connect-timeout").getSeconds, TimeUnit.SECONDS)
-
-  override val socketConnectTimeout = FiniteDuration(config.getDuration("socket-connect-timeout").getSeconds, TimeUnit.SECONDS)
-
-  override val openBucketRetryTimeout = FiniteDuration(config.getDuration("open-bucket-retry-timeout").getSeconds, TimeUnit.SECONDS)
-}
-
-object CouchbaseEnvironmentConfig {
-  def apply(system: ActorSystem) = {
-    new DefaultCouchbaseEnvironmentConfig(system.settings.config.getConfig("couchbase-environment"))
   }
 }
